requirements.txt
streamlit
azure-cognitiveservices-vision-computervision
python-dotenv
Levenshtein
Pillow